refactor: ♻️ refactor action panel section
