+++ apps/builder/src/page/App/components/Actions/ActionGuidePanel/components/ActionSection/interface.ts
@@ -1,0 +1,9 @@
+import { ActionType } from "@/redux/currentApp/action/actionState"
+
+export interface PanelSectionProps {
+  actionTypes: ActionType[]
+  changeLoading: (isLoading: boolean) => void
+  filterFunc?: (actionType: ActionType) => boolean
+  title: string
+  hasMore: boolean
+}
