+++ apps/builder/src/page/App/components/Actions/ActionGuidePanel/constans.ts
@@ -1,6 +1,7 @@
+import { isCloudVersion } from "@illa-public/utils"
 import { ActionType } from "@/redux/currentApp/action/actionState"
 
-export const RECOMMEND_RESOURCES: ActionType[] = [
+export const RECOMMEND_RESOURCES_SELF_HOST: ActionType[] = [
   "restapi",
   "mysql",
   "postgresql",
@@ -8,11 +9,16 @@ export const RECOMMEND_RESOURCES: ActionType[] = [
   "supabasedb",
 ]
 
-export const MORE_DATA_TYPE: ActionType[] = [
-  "aiagent",
+export const MORE_DATA_TYPE_SELF_HOST: ActionType[] = [
   "transformer",
   "globalData",
-  "illadrive",
 ]
 
 export const ONLY_CLOUD_MODE_DATA_TYPE: ActionType[] = ["aiagent", "illadrive"]
+
+export const RECOMMEND_RESOURCES_CLOUD: ActionType[] = ["restapi", "mysql"]
+
+export function filterSelfHostActionType(actionType: ActionType) {
+  if (!isCloudVersion) return !ONLY_CLOUD_MODE_DATA_TYPE.includes(actionType)
+  return !!actionType
+}
