+++ apps/builder/src/page/App/components/Actions/ActionGuidePanel/style.ts
@@ -16,65 +16,6 @@ export const guidePanelContainerStyle = css`
   margin: 0 auto;
 `
 
-export const categoryTitleStyle = css`
-  padding: 0;
-  padding-top: 16px;
-  padding-bottom: 8px;
-  color: ${getColor("grayBlue", "04")};
-  font-size: 14px;
-  font-weight: 500;
-  line-height: 22px;
-  margin: 0;
-  width: 552px;
-`
-
-export const categoryItemContainerStyle = css`
-  display: grid;
-  gap: 16px;
-  grid-template-columns: repeat(3, 173px);
-`
-
-export const basicButtonStyle = css`
-  display: flex;
-  width: 100%;
-  padding: 8px 16px;
-  gap: 12px;
-  align-items: center;
-  border: 1px solid ${getColor("grayBlue", "08")};
-  border-radius: 8px;
-  background-color: white;
-  cursor: pointer;
-  transition: all 0.2s ease-in-out;
-  &:hover {
-    box-shadow: 0 4px 10px 0 ${getColor("blackAlpha", "07")};
-    background-color: ${getColor("techPurple", "07")};
-    border-color: ${getColor("techPurple", "01")};
-  }
-`
-
-export const iconHotSpot = css`
-  font-size: 16px;
-  display: flex;
-  flex: none;
-  justify-content: center;
-  align-items: center;
-  color: ${getColor("techPurple", "01")};
-`
-
-export const moreTipsStyle = css`
-  font-weight: 500;
-  font-size: 12px;
-  line-height: 20px;
-  color: ${getColor("techPurple", "01")};
-`
-
-export const categoryItemNameStyle = css`
-  font-weight: 500;
-  font-size: 12px;
-  line-height: 20px;
-  color: ${getColor("grayBlue", "02")};
-`
-
 export const loadingContainerStyle = css`
   width: 100%;
   height: 100%;
